agentic/requirements.txt
# yousef_shtiwe Agent Dependencies

# LangChain core
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
langgraph>=0.2.0
langchain-core>=0.3.0

# AWS Bedrock integration
langchain-aws>=0.2.0
boto3>=1.35.0

# MCP integration
langchain-mcp-adapters>=0.1.0

# Neo4j integration
langchain-community>=0.3.0
langchain-neo4j>=0.1.0
neo4j>=5.0.0

# Tavily web search
langchain-tavily>=0.1.0

# Vector knowledge base (heavy deps in requirements-kb.txt)
pyyaml>=6.0
openai>=1.0.0

# FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0

# Async support
anyio>=4.0.0

# WebSocket client for kali terminal proxy
websockets>=12.0

# CypherFix CodeFix agent
PyGithub>=2.1.0
gitpython>=3.1.0

# Tree-sitter navigation tools (symbols, find_definition, find_references, repo_map)
tree-sitter==0.21.3
tree-sitter-languages==1.8.0
networkx>=3.0
